requirements.txt
Flask==2.1.1
sqlalchemy==1.4.35
flask_marshmallow==0.14.0
psycopg2-binary==2.9.3
flask_sqlalchemy==2.5.1
marshmallow-sqlalchemy==0.28.0
redis==4.2.2
pyjwt==2.3.0